+++ test/renderer-plugins-test.js
@@ -1,9 +1,9 @@
 import test from 'tape'
-import render from '../src/index.js'
+import { Arcdown } from '../src/index.js'
 
 test('renderer plugin overrides', async (t) => {
   const TOC_CLASS = 'pageToC'
-  const file = /* md */`
+  const file = /* md */ `
 ## Deploy to AWS
 
 [AWS](https://aws.amazon.com/) is a cloud computing platform that makes it easy to build, deploy, and manage applications and services.
@@ -11,7 +11,8 @@ test('renderer plugin overrides', async (t) => {
 
   const options = {
     pluginOverrides: {
-      markdownItTocAndAnchor: { tocClassName: TOC_CLASS },
+      markdownItAnchor: { tabIndex: 42 },
+      markdownItToc: { containerClass: TOC_CLASS },
       markdownItClass: {
         h2: [ 'title' ],
         p: [ 'prose' ],
@@ -20,13 +21,12 @@ test('renderer plugin overrides', async (t) => {
     },
   }
 
-  const {
-    html,
-    tocHtml,
-  } = await render(file, options)
+  const renderer = new Arcdown(options)
+  const { html, tocHtml } = await renderer.render(file)
 
   t.ok(tocHtml.indexOf(`class="${TOC_CLASS}`) >= 0, 'ToC class is present')
-  t.ok(html.indexOf('target="_blank">AWS</a>') < 0, 'External link targets = blank')
+  t.ok(html.indexOf('target="_blank"') < 0, 'External link targets != blank')
+  t.ok(html.indexOf('tabindex="42"') > 0, 'Heading anchor has tabindex')
   t.ok(html.indexOf('<h2 class="title"') >= 0, 'h2.title')
   t.ok(html.indexOf('<p class="prose"') >= 0, 'p.prose')