+++ types/index.d.ts
@@ -1,10 +1,13 @@
 /// <reference types="node" />
 
+import MarkdownIt = require('markdown-it/lib');
+
 // TODO: types for each default plugin config
 export interface DefaultPlugins {
   markdownItClass?: object | boolean;
   markdownItExternalAnchor?: object | boolean;
-  markdownItTocAndAnchor?: object | boolean;
+  markdownItAnchor?: object | boolean;
+  markdownItToc?: object | boolean;
 }
 
 export interface HljsOptions {
@@ -19,20 +22,8 @@ export interface RendererOptions {
   hljs?: HljsOptions;
   pluginOverrides?: DefaultPlugins;
   plugins?: object;
-  renderer?: {
-    render: (markdown: string) => string;
-    use?: any;
-  };
+  renderer?: MarkdownIt;
 }
-
-export const defaultPlugins: DefaultPlugins;
-
-export function slugify(s: string): string;
-
-export function createHighlight(
-  options: object,
-  foundLanguages: string[] | null,
-): Promise<(code: any, language: any) => string>;
 
 export interface RenderResult {
   html: string;
@@ -42,7 +33,9 @@ export interface RenderResult {
   frontmatter?: Record<string, any>;
 }
 
-export default function render(
-  mdFile: Buffer | string,
-  rendererOptions?: RendererOptions,
-): Promise<RenderResult>;
+export class Arcdown {
+  constructor(options: RendererOptions);
+  static slugify(s: string): string;
+  static findLanguages(mdContent: string): Set<string>;
+  render(mdFile: Buffer | string): Promise<RenderResult>;
+}